third_party/WebKit/LayoutTests/editing/input/style-change-during-input-expected.txt

The text field should accept typed characters.
PASS window.target.value is "hello"
PASS successfullyParsed is true

TEST COMPLETE
